requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
gunicorn
psycopg2-binary
